--- src/1-basic-types/nishiyama/namespaces/Flyable.ts
@@ -1,7 +1,0 @@
-namespace Capability {
-  export class Flyable implements CapabilityCheker {
-    isCapable(species: string): boolean {
-      return species === 'bird';
-    }
-  }
-}
