--- src/1-basic-types/nishiyama/namespaces/Healable.ts
@@ -1,7 +1,0 @@
-namespace Capability {
-  export class Healable implements CapabilityCheker {
-    isCapable(species: string): boolean {
-      return species === 'cat' || species === 'human';
-    }
-  }
-}
