--- src/1-basic-types/nishiyama/namespaces/Speakable.ts
@@ -1,7 +1,0 @@
-namespace Capability {
-  export class Speakable implements CapabilityCheker {
-    isCapable(species: string): boolean {
-      return species === 'human';
-    }
-  }
-}
