--- src/1-basic-types/nishiyama/namespaces/Swimmable.ts
@@ -1,7 +1,0 @@
-namespace Capability {
-  export class Swimmable implements CapabilityCheker {
-    isCapable(species: string): boolean {
-      return species === 'fish';
-    }
-  }
-}
